+++ server/util.go
@@ -354,8 +354,7 @@ func AddCommands[T types.Application](rootCmd *cobra.Command, appCreator types.A
 		startCmd,
 		cometCmd,
 		version.NewVersionCommand(),
-		NewRollbackCmd(appCreator),
-		ModuleHashByHeightQuery(appCreator),
+		NewRollbackCmd(defaultNodeHome),
 	)
 }